Large update to read_authors
This attempts to introduce an AI algorithm to more smartly match authors.
As opposed to the previous iteration, this technique matches against all authors at once, instead of building a novel author list slowly. The advantages of this are we can use the collective information of each entry (RI, OI, email, university, etc). It appears relatively faster than the previous iteration as well. The rules for matching are as follows, and as usually is hierarchical:
1. Matches all OIs first
2. Matches RIs from remaining entries
3. Matches by email from remaining entries
4. Matches by entire name (middle must be included), from remaining entries.

The final set of entries is subsetted for further matching. This subset must have either a middle initial, a universitiy affiliation, or an email address. Otherwise we assume that the name is unmatchable and also cant be matched against. Further iterations might have address matching.

This final subset of entries is matched against all entries in the subset that match the last name and first name PLUS one of the following criteria:
1. Middle name matches
2. University matches
3. Middie initial matches (if only an initial)

If no match exists after this jaro_winkler is run against this subset of names.